Added non AABB rectangular hitboxes and added this hitbox to the obstacle placement ability

+++ src/game_state.h
@@ -31,7 +31,7 @@ struct GameState {
     union {
         struct
         {
-            i32 rotation = 0;
+            float rotation = 0;
         } obstacleAbility;
 
         struct
